requirements.txt
pandas==1.4.3
xlrd==2.0.1
Pillow==9.2.0
plotly==5.9.0
streamlit==1.11.1
streamlit_lottie==0.0.3
streamlit_option_menu==0.3.2
